Fixed warnings in GCamp.cpp

+++ Goblin Camp/src/GCamp.cpp
@@ -312,7 +312,7 @@ void LoadMenu() {
 		TCODConsole::root->print(edgex + (width / 2), edgey + 1, "ESC to cancel.");
 		TCODConsole::root->setAlignment(TCOD_LEFT);
 
-		for (int i = 0; i < list.size(); ++i) {
+		for (unsigned int i = 0; i < list.size(); ++i) {
 			if (selected == i) {
 				TCODConsole::root->setDefaultForeground(TCODColor::black);
 				TCODConsole::root->setDefaultBackground(TCODColor::white);
@@ -352,7 +352,7 @@ void LoadMenu() {
 		if (!mouseStatus.lbutton && lButtonDown) {
 			lButtonDown = false;
 			
-			if (selected < list.size() && selected >= 0) {
+			if (selected < (signed int)list.size() && selected >= 0) {
 				Data::LoadGame(list[selected].filename);
 				MainLoop();
 				break;
